setting states to be specified in prefabs rather than code
