+++ Assets/_Scripts/Level/Brains/CharacterStateConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e29462426a12b4e0bb7bcbb51ad4f715
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
